pyproject.toml
[project]
name = "test-claude-api-key"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.84.0",
    "python-dotenv>=1.2.1",
    "pdfplumber>=0.10",
    "langchain>=0.2",
    "langchain-community>=0.2",
    "sentence-transformers>=2.7",
    "torch>=2.0",
    "chromadb>=0.5",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]
